# Allow running over empty Nano files