--- a/columnflow/tasks/reduction.py
+++ b/columnflow/tasks/reduction.py
@@ -488,8 +488,7 @@ class MergeReducedEvents(
         with law.localize_file_targets(
             inputs, cache=False, mode="r",
         ) as localized_inputs:
-            from IPython import embed
-            embed(header="merging inputs in MergeReducedEvents")
+            
             metadata = [pq.read_metadata(inp.abspath) for inp in localized_inputs]
 
             # empty nano files are not considered for merging