There are some NanoAOD files (*) with 0 events that fail when running through our workflow. This PR fixes that by skipping these files when running calibrators/selectors or when merging files.
Additionally, I also added a check for inconsistencies in number of fields when merging ReduceEvents outputs.



(*) example: file 0 from `data_muoneg_f`in the run3_2022_postEE_nano_v12 campaign:
```
# DAS
/MuonEG/Run2022F-22Sep2023-v1/NANOAOD
```
```
# file
/store/data/Run2022F/MuonEG/NANOAOD/22Sep2023-v1/50000/4d76213a-ef14-411a-9558-559a6df3f978.root
```