@pkausw 2024-08-12
Thanks for this PR! @mafrahm and I had an offline chat about this project, and came up with a couple of things that should be checked before merging (collecting it here to keep track)
- [ ] does the check of the meta data work with remote files in the merging step?
- [ ] how costly would it be to perform a check for empty files already in the GetDatasetLFNs step?
- [ ] how costly would it be to compare the (sets of) names of the columns in the files to merge instead of the number of columns? 

Let me know if I forgot a check!