@pkausw 2024-08-13
>     how costly would it be to perform a check for empty files already in the GetDatasetLFNs step?

Update: I have tested how much time it would take to directly check the number of events based on the dasgoclient. For this, I modified the `get_dataset_lfns` function:

```python
def get_dataset_lfns_filter_nevents(
            dataset_inst: od.Dataset,
            shift_inst: od.Shift,
            dataset_key: str,
        ) -> list[str]:
            """
            Get the LNF information with the ``dasgoclient``.

            :param dataset_inst: Current dataset instance, currently not used.
            :param shift_inst: Current shift instance, currently not used.
            :param dataset_key: DAS key identifier for the current dataset.
            :raises Exception: If query with ``dasgoclient`` fails.
            :return: The list of LFNs corresponding to the dataset with the identifier *dataset_key*.
            """
            import subprocess
            import json
            code, out, _ = law.util.interruptable_popen(
                f"dasgoclient --query='file dataset={dataset_key}' -json",
                shell=True,
                stdout=subprocess.PIPE,
                executable="/bin/bash",
            )
            if code != 0:
                raise Exception(f"dasgoclient query failed:\n{out}")

            das_info = json.loads(out.strip())
            valid_files = list(filter(
                lambda x: x["file"][0]["is_file_valid"]==1 and x["file"][0]["nevents"] >0, das_info
            ))
            return [
                x["file"][0]["name"]
                for x in valid_files
            ]
 ```
 
 I did a small-scale timing study with and without this function using the command
 
 ```bash
 time law run cf.GetDatasetLFNs --config run3_2022_postEE_cms_v12 --dataset data_muoneg_f --remove-output 0,a,y
 ```
 
 5 times. Here are the results:
 
 | | time (original) | time (nevents check) |
 | --- | --- | --- |
 | real | 21.1396 +- 0.607 | 11.8054 +- 0.3181 |
 | user | 24.4952 +- 0.43 | 12.8302 +- 0.3898 |
 | sys |  1.6606 +- 0.0832 | 0.6584 +- 0.0123 | 
 
 
 These values have to be taken with a grain of salt I guess - e.g. I don't really understand why the query becomes faster. Bottom line though is that performing this check doesn't seem to take longer, which would have the nice side effect that we don't need to consider the empty files in the branch map to begin with. 
 
 This of course would mean that we need to perform such a filter already on cmsdb level, since the files that the `dasgoclient` query delivers doesn't match with the number of files specified in the corresponding cmsdb campaign. I had a short offline discussion with @riga, and we think it would be nice to add broken/empty files as auxiliary information to the individual datasets. This way, we would be able to track these files and update the cmsdb entry when/if these issues are resolved in the future. For the dataset in question, this could look something like this
 
 ```python
 cpn.add_dataset(
    name="data_muoneg_f",
    id=14784482,
    is_data=True,
    processes=[procs.data_muoneg],
    keys=[
        "/MuonEG/Run2022F-22Sep2023-v1/NANOAOD",  # noqa
    ],
    n_files=95,
    n_events=38219969,
    aux={
        "era": "F",
        "broken_files": [
            '/store/data/Run2022F/MuonEG/NANOAOD/22Sep2023-v1/50000/4d76213a-ef14-411a-9558-559a6df3f978.root',
            '/store/data/Run2022F/MuonEG/NANOAOD/22Sep2023-v1/50000/4fb72196-3b02-4499-8f6c-a54e15692b32.root'
        ]
    },
)
```

In this example, the number of files to process would be `n_files - len(broken_files`. What do you think @mafrahm ?