@pkausw 2024-08-16
Update on this item:

>     * [x]  does the check of the meta data work with remote files in the merging step?

-> no, it doesn't. I have pushed a fix that first localizes first the inputs and the output of the task, tested with CERN eos as remote target