@pkausw 2024-08-16
I looked into the final point on the list @mafrahm . I implemented an explicit check for the column names and ran the MergeReducedEvents task over all 93 files of the above mentioned dataset. The timings I took indicate that there isn't a large run time difference between the versions - indeed, the version of the code with the new check seems to run ~10s faster on average than the current implementation (though the statistics is very limited of course).

With this in mind, I don't see a reason why we shouldn't check this. If you agree, I can push my implementation, but I'm also happy to discuss of course :+1: